+++ src/Tetrahedron.cpp
@@ -2,7 +2,7 @@
 
 void Tetrahedron::init() {
     // Calculate the center of the tetrahedron
-    center = Eigen::Vector3d::Zero();
+    center = Eigen::Vector3f::Zero();
     for (int i = 0; i < 4; i++) {
         center += vertices[i];
     }
@@ -17,10 +17,10 @@ void Tetrahedron::init() {
     invTransposeMatrix = edgeMatrix.inverse().transpose();
 }
 
-int Tetrahedron::contains(const Eigen::Vector3d& p) const {
+int Tetrahedron::contains(const Eigen::Vector3f& p) const {
     // having base point A, inverseTrasnpose * AP 
-    Eigen::Vector3d AP = p - vertices[0];
-    Eigen::Vector3d bary = invTransposeMatrix * AP;
+    Eigen::Vector3f AP = p - vertices[0];
+    Eigen::Vector3f bary = invTransposeMatrix * AP;
     float d = std::max(std::max(-bary[0], -bary[1]), std::max(-bary[2], bary[0]+ bary[1]+ bary[2] - 1));
     if (d > 0) {
         // calculate norm2 of center to p
